+++ tests/testthat/test_expr_ALL.R
@@ -3,7 +3,7 @@ context("Gene expression data (ALL)")
 test_that("Vanilla tests for ALL data", {
     data("expr_ALL")
     dat <- expr_ALL
-    expect_equal(nrow(dat), 12625)
+    expect_equal(nrow(dat), 9038)
     expect_equal(ncol(dat), 79)
     cn <- colnames(dat)
     tab <- table(cn)
@@ -13,12 +13,12 @@ test_that("Vanilla tests for ALL data", {
 })
 
 
-test_that("Tests for ALL probe annotation data", {
+test_that("Tests for expr_ALL gene annotation data", {
     data("expr_ALL")
     data("expr_ALL_annotation")
 
     ids <- rownames(expr_ALL)
-    ann_ids <- expr_ALL_annotation[["affy_hg_u95av2"]]
+    ann_ids <- expr_ALL_annotation[["ensembl_gene_id"]]
     expect_true(all(ann_ids %in% ids))
 
     chr <- expr_ALL_annotation[["chromosome_name"]]
@@ -28,4 +28,19 @@ test_that("Tests for ALL probe annotation data", {
     spos <- expr_ALL_annotation[["start_position"]]
     epos <- expr_ALL_annotation[["end_position"]]
     expect_true(all(epos > spos))
+})
+
+
+test_that("Tests for GO annotation data", {
+    data("expr_ALL")
+    data("expr_ALL_GO")
+
+    expect_equal(nrow(expr_ALL_GO), 9038)
+    expect_equal(ncol(expr_ALL_GO), 144)
+
+    ids <- rownames(expr_ALL)
+    ann_ids <- rownames(expr_ALL_GO)
+    expect_true(identical(ids, ann_ids))
+
+    expect_true(all(expr_ALL_GO %in% c(0,1)))
 })