+++ NEWS.md
@@ -1,3 +1,16 @@
+# sansSouci.data 0.3.0
+
+* Updated 'expr_ALL': now *gene*-level instead of probe set level. 
+* Updated 'expr_ALL': row names are now ENSEMBL gene ids.
+* Updated 'expr_ALL_annotation' according to the above updates. 
+* Added 'expr_ALL_GO': GO annotation data associated to expression data set.
+* Updated documentation and tests accordingly.
+
+# sansSouci.data 0.2.2
+
+* Added 'hgu95av2_GO_BP': GO annotation data for chip type "hgu95av2".
+* Now only 4 decimal places for 'fMRI_localizer' data set.
+
 # sansSouci.data 0.2.1
 
 * Added probe annotation data for differential expression data set.